Refactor topic-link type and add resource page
- Refactor the topic-link type to fix a typo in the file name and update the export statement.
- Add a new resource page with functionality to search, play audio, and display different file types.
