--- src/app/resource/page.tsx
@@ -1,5 +1,0 @@
-import { Resource } from '@/components/Resource/Resource'
-
-export default async function Page() {
-  return <Resource />
-}
